fix:ajustes gerais

+++ src/Infrastructure/Repositories/ProductRepository.ts
@@ -3,6 +3,9 @@ import type { ProductEntity } from "../../Data/Db/Entities/Product";
 import type { IProductRepository } from "../Interfaces/IProductRepository";
 
 export class ProductRepository implements IProductRepository {
+  findManyByIds(ids: string[]): Promise<ProductEntity[]> {
+    throw new Error("Method not implemented.");
+  }
   async create(productData: ProductEntity): Promise<ProductEntity> {
     const newProduct = await prisma.product.create({
       data: {
@@ -69,24 +72,58 @@ export class ProductRepository implements IProductRepository {
 
   async delete(slug: string): Promise<boolean> {
     try {
+      const product = await prisma.product.findUnique({
+        where: { slug: slug },
+        select: { id: true },
+      });
+
+      if (!product) {
+        // Produto nem existe, então não há o que deletar
+        return false;
+      }
+
+      // Passo 2: Contar quantos pedidos estão associados a este produto
+      const orderCount = await prisma.order.count({
+        where: {
+          products: {
+            some: {
+              productId: product.id,
+            },
+          },
+        },
+      });
+
+      if (orderCount > 0) {
+        console.error(
+          `Tentativa de deletar produto ${slug} que possui ${orderCount} pedidos.`
+        );
+        return false;
+      }
+
       await prisma.product.delete({
         where: { slug: slug },
       });
+
       return true;
     } catch (error) {
+      console.error("Erro ao deletar produto:", error);
       return false;
     }
   }
 
-  async findManyByIds(ids: string[]): Promise<Product[]> {
-    return prisma.product.findMany({
-      where: {
-        id: {
-          in: ids,
-        },
-      },
-    });
-  }
+  // async findManyByIds(ids: string[]): Promise<ProductEntity[]> {
+  //   const products = await prisma.product.findMany({
+  //     where: {
+  //       id: {
+  //         in: ids,
+  //       },
+  //     },
+  //   });
+
+  //   return products.map((product) => ({
+  //     ...product,
+  //     createdAt: product.createdAt || new Date(),
+  //     generateSlug: () => product.slug,
+  //   })) as ProductEntity[];
+  // }
 }
-
-
